foi realizado o refactory nos serializadores assim como foi feito a adequação onde ele era utilizado.
closes #25
