+++ api.http
@@ -13,8 +13,8 @@ POST {{baseURL}}/api/states
 Content-Type: application/json
 
 {
-    "name": "Distrito federal",
-    "abbreviation": "DF"
+    "name": "São Paulo",
+    "abbreviation": "SP"
 }
