+++ tracker/views.py
@@ -1,5 +1,5 @@
 from helpers import restfy
-from .serialization import StateSerializer, CitySerializer
+from .serializers import StateSerializer, CitySerializer
 
 
 state_index, state_by_id = restfy.make_rest(StateSerializer)
